new update

+++ application/views/pages_examples/combos_table.php
@@ -39,37 +39,40 @@
         </div>
     </div>
     <div class="row">
-        <?php foreach ($response as $row) { ?>
-            <div class="col-md-4 contem" style="margin-top: 20px">
-                <div class="card">
-                    <div class="card-body">
-                        <div class="row">
-                            <div class="col-md-9">
-                                <b class="card-title"><?php echo $row['establishment_id'] ?></b>
-                                <p class="card-text">
-                                    <?php echo $row['description'] ?><br>
-                                    &nbsp;R$ <?php echo $row['value'] ?>
+        <?php
+        if ($response) {
+            foreach ($response['combos'] as $row) { ?>
+                <div class="col-md-4 contem" style="margin-top: 20px">
+                    <div class="card">
+                        <div class="card-body">
+                            <div class="row">
+                                <div class="col-md-9">
+                                    <b class="card-title"><?php echo $row['establishment_id'] ?></b>
+                                    <p class="card-text">
+                                        <?php echo $row['description'] ?><br>
+                                        &nbsp;R$ <?php echo $row['value'] ?>
 
-                                </p>
+                                    </p>
+                                </div>
+                                <div class="col-md-1">
+                                    <a href="<?php echo base_url() ?>Combos/new_user/<?php echo $row['id'] ?>"><i
+                                                class="fas fa-pencil-alt"
+                                                style="color: gray"></i></a>
+                                </div>
+                                <div class="col-md-1">
+                                    <a href="<?php echo base_url() ?>Combos/delete/<?php echo $row['id'] ?>"> <i
+                                                class="fas fa-times"
+                                                style="color: gray"></i></a>
+                                </div>
                             </div>
-                            <div class="col-md-1">
-                                <a href="<?php echo base_url() ?>Combos/new_user/<?php echo $row['id'] ?>"><i
-                                            class="fas fa-pencil-alt"
-                                            style="color: gray"></i></a>
-                            </div>
-                            <div class="col-md-1">
-                                <a href="<?php echo base_url() ?>Combos/delete/<?php echo $row['id'] ?>"> <i
-                                            class="fas fa-times"
-                                            style="color: gray"></i></a>
-                            </div>
+                            <!--
+
+                            <!--                    <a href="#" class="btn btn-primary">Go somewhere</a>-->
                         </div>
-                        <!--
-
-                        <!--                    <a href="#" class="btn btn-primary">Go somewhere</a>-->
                     </div>
                 </div>
-            </div>
-        <?php } ?>
+            <?php }
+        } ?>
     </div>
     <br>
     <div class="row">
